predictor/requirements.txt
kafka-python-ng==2.2.2
redis==5.0.3
